% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStopAnova.R
\name{OneStopAnova}
\alias{OneStopAnova}
\title{OneStopAnova}
\usage{
OneStopAnova(Quantitative = "",
             Qualitative = "",
             Qualitative2 = "",
             var_names = c(Quantitative = "",
                           Qualitative = "",
                           Qualitative2 = ""),
             Log = TRUE/FALSE,
             Tukey = TRUE/FALSE)
}
\arguments{
\item{Quantitative:}{The Quantitative variable}

\item{Qualitative:}{The Qualitative variable}

\item{Qualitative2:}{The Second Qualitative variable}

\item{var_names:}{The Names of the Selected Variables}

\item{Log:}{FALSE by default. If TRUE, Log Transformation of the Quantitative Variable}

\item{Tukey:}{FALSE by default. If TRUE, Tukey Test on the Selected Variables}
}
\description{
This package generates One-Way and Two-Way Anova with assumptions verification, optional Log and optional Tukey.
}
\examples{
OneStopAnova(Quantitative = "calcium$Concentration",
             Qualitative = "calcium$Trait",
             Qualitative2 = "calcium$Sexe",
             var_names = c(Quantitative = "Concentration",
                           Qualitative = "Trait",
                           Qualitative2 = "Sexe"),
             Log = TRUE,
             Tukey = TRUE)

}
