% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStopAnova_IC.R
\name{OneStopAnova_IC}
\alias{OneStopAnova_IC}
\title{OneStopAnova_IC}
\usage{
OneStopAnova_IC(aov = "", p = "")
}
\arguments{
\item{aov:}{object holding the anova results}

\item{p:}{confidence level (0.025 = 95 percent)}
}
\description{
This function generates Confidence Interval from One-Way | Two-Way Anova function object
}
\examples{
aov = aov(Concentration ~ Trait + Sexe + Trait:Sexe, data = calcium)
OneStopAnova_IC(aov = aov, p = 0.025)

}
